more refactoring to reduce code smell/duplication
Two tests are wrong when considering a clean SUT.
The @Before method now adds a Customer and an Event, thus the tests
`shouldListAllEvents` and `shouldListAllCustomers` are responsible to
add the initially created Customer and Event to the test list.

I think, the expected and more correct way would be add a new test rule
that allows bypassing the `@Before` method to avoid this problem.