implement real mocking of the blacklist
Previously, the test just checked if the `Blacklist` returns `false`
or `true`, but not if the booking will be rejected, as there was no
rejection programmed in the service.

Checking with the example made by Prof. Dr. Garmann, I think this
should be the right solution now.

The `Blacklist` is an interface, there is a stubborn and basic
`BlacklistImpl` which currently just returns false, to ensure that
`Services` class is testable and doesn't force all other tests to
mock the `Blacklist` too. Hopefully this makes sense.